requirements.txt
Flask==2.3.2
PyYAML>=6.0
